feat(extracttriggers): function name param to allow multi-trigger files
The extractTrigger function is now passed an additional param that indicates the current function
name (if not deploying). In conjunction with customising funcNameFromRelPath, this could be used to
allow a setup where more than one trigger per file is stored. This isn't the best for performance
but we have decided to allow the flexibility.

feat #6