feat(default.ts): remove deprecated default export function
Initial versions of this package provided its utility function as a default export of the module.
This is no longer the case and this function was deprecated. This commit deletes the export from the
module.

BREAKING CHANGE: The breaking change with this update is that any codebases utilising the default
export will no longer work. Please update to use the exported named function exportFunctions.