feat(exportfunctions): update search glob to sensible default to find both js and ts files
This will enable exportFunctions to work both in a js environment and ts environment|This is usefull
because when serving applications they are often run with ts-node or otherwise not compiled to js,
but when deployed are compiled to js. Meaning the same code doesn't work between environments. This
should fix that.

BREAKING CHANGE: The new default search glob may break some setups in rare situations where js files
and ts files are in the same directory.